pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pax"
version = "0.2.0"
description = "Perception-aware corridor analytics"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Gil Raitses" }
]
dependencies = [
  "requests>=2.32",
  "pydantic>=2.6",
  "pydantic-settings>=2.4",
  "numpy>=1.26",
  "pandas>=2.1",
  "geopandas>=0.14",
  "shapely>=2.0",
  "scipy>=1.11",
  "networkx>=3.2",
  "scikit-learn>=1.4",
  "tqdm>=4.66",
  "python-dotenv>=1.0",
  "pyarrow>=15.0",
  "google-cloud-storage>=2.16",
  "matplotlib>=3.8",
  "pillow>=10.3",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
